Passing data as prop to Blog
