+++ pages/index.vue
@@ -1,4 +1,4 @@
 <template>
-  <p>Redirecting...</p>
+  <blog :page="page"/>
 </template>
 <script src="./index.js"></script>
